Refactor inode module
